Macro constify::constify

macro_rules! constify {
    {
        $(
        const $const_var:ident: $const_ty:ty = $match_val:expr => $($const_expr:expr),+;
        )*

        $return:expr
    } => { ... };
}

Convert runtime values to consts.

This macro compares runtime expressions to the provided constants and evaluates the given expression with the matching constants. All cases must be covered; to match only a subset of values use try_constify.

The expressions are evaluated in the order they are provided.

An example:

fn sum_impl<const A: bool, const B: bool>(a: u32, b: u32) -> u32 {
    let mut sum = 0;
    if A {
        sum += a;
    }
    if B {
        sum += b;
    }
    sum
}

fn sum(a: u32, b: u32, add_a: bool, add_b: bool) -> u32 {
    constify::constify! (
        const A: bool = add_a => true, false;
        const B: bool = add_b => true, false;

        sum_impl::<A, B>(a, b)
    )
}

assert_eq!(sum(3, 4, false, false), 0);
assert_eq!(sum(3, 4, true, false), 3);
assert_eq!(sum(3, 4, false, true), 4);
assert_eq!(sum(3, 4, true, true), 7);